pyproject.toml
[tool.poetry]
name = "dash-router-examples"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
dash-mantine-components = "1.2.0"
dash-iconify = "^0.1.2"
orjson = "^3.10.15"
dash = "3.0.4"
pandas = "^2.2.3"
dash-ag-grid = "v31.3.1"
dash-aggrid-scales = "^0.2.1"
dash-dynamic-grid-layout = "^0.1.2"
dash-extensions = "2.0.0"
aiocache = "^0.12.3"
plotly = "6.0.1"
nbformat = "^5.10.4"
dash-resizable-panels = "^0.1.0"
dash-snap-grid = "^0.1.0"
sqlalchemy = "^2.0.41"
redis = "^6.1.0"
python-dotenv = "^1.1.0"
asyncpg = "^0.30.0"
uvicorn = "^0.34.2"
flash-router = "^0.3.0b2"
dash-flash = "^1.0.0"
greenlet = "^3.2.2"
kaleido = "0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
